Add files via upload
First commit of all the files I'm using right now. The LCD is just how to initialize the LCD screen. The temp_sensor code is the first run of the main code but threw exceptions when it occasionally couldn't reach the digital sensor. temp_sensor_error adding in some exception handling to get around that. temp_slack alerts the slack channel and will hopefully also text when temperature is too high.